requirements.txt
streamlit==1.38.0
pandas
matplotlib
python-dotenv
pandasai
